+++ WheelWizard/Views/Pages/Settings/AppInfo.axaml
@@ -134,6 +134,14 @@
                     </StackPanel>
                 </Grid>
             </Border>
+            
+         
+        </StackPanel>
+        
+        <StackPanel HorizontalAlignment="Right" VerticalAlignment="Bottom">
+            <TextBlock x:Name="ReleaseText" HorizontalAlignment="Right" Classes="TinyText" Text="DEV BUILD"/>
+            <TextBlock x:Name="WhWzVersionText" HorizontalAlignment="Right" Classes="TinyText" Text="Example WhWz version 1.0.0" />
+            <TextBlock x:Name="RrVersionText" HorizontalAlignment="Right" Classes="TinyText" Text="Example RR version 1.0.0" />
         </StackPanel>
     </Grid>
 </UserControl>
