+++ WheelWizard/Views/Pages/Settings/SettingsPage.axaml.cs
@@ -13,41 +13,15 @@ public partial class SettingsPage : UserControlBase
     public SettingsPage()
         : this(new WhWzSettings()) { }
 
-    [Inject]
-    private ICustomDistributionSingletonService CustomDistributionSingletonService { get; set; } = null!;
-
     public SettingsPage(UserControl initialSettingsPage)
     {
         InitializeComponent();
 
-        RrVersionText.Text = "RR: " + CustomDistributionSingletonService.RetroRewind.GetCurrentVersion();
-
-        var part1 = "Release";
-        var part2 = "Unknown OS";
 #if DEBUG
-        part1 = "Dev";
         DevButton.IsVisible = true;
 #endif
-        // We intentionally use preprocessor directives (#if, #elif, #endif) instead of Environment.OSVersion
-        // because 'part2' represents the OS this code was built for, not the OS it is currently running on.
-#if WINDOWS
-        part2 = "Windows";
-#elif LINUX
-        part2 = "Linux";
-#elif MACOS
-        part2 = "Macos";
-#endif
 
-        ReleaseText.Text = $"{part1} - {part2}";
         SettingsContent.Content = initialSettingsPage;
-    }
-
-    protected override void OnInitialized()
-    {
-        var branding = App.Services.GetRequiredService<IBrandingSingletonService>().Branding;
-        WhWzVersionText.Text = $"WhWz: v{branding.Version}";
-
-        base.OnInitialized();
     }
 
     private void TopBarRadio_OnClick(object? sender, RoutedEventArgs e)
